Stop Addresses and links of runs from being published
The addresses of where people are running are being put into 
a public irc channel. In some countries the first part of a post code
can be a small as a couple miles. The Strava link has also been removed
as this is another area where peoples privacy can be compromised.